Replace `_maybetail` with `safe_tail`
This was motivated by the disappearance of `_maybetail` from `Base`,
which causes `AxisKeys` to not be compatible with v1.9. As it happens,
`safe_tail` is identical to `_maybetail`, thus, it would seem that the
latter was eliminated due to redundancy. Inspection of the v1.6
branch of JuliaLang confirms that `safe_tail` exists in its same form,
thus, given the `[compat]` for this `AxisKeys`, this seems like a
harmless change which enables compatibility with `master`.